+++ src/zachfix/main_exe.cpp
@@ -27,6 +27,7 @@ constexpr DpBuildProfile kSteam101bProfile{
     0x00309C40,
     0x001E6D40,
     0x001EC37F,
+    0x002C6296,
     0x002E15F0,
     0x00335BB0,
     0x0014C1D1
@@ -47,6 +48,7 @@ constexpr DpBuildProfile kGog101bProfile{
     0x00309BA0,
     0x001E6E10,
     0x001EC44F,
+    0x002C5D96,
     0x002E1630,
     0x003358C0,
     0x0014C2A1